⚡️ perf(speed, acceleration): 调整速度和加速度相关的数值类型和限制
- 将 `QSpinBox` 更改为 `QDoubleSpinBox` 以支持更精细的数值输入。
- 更新了 `config/settings.json` 中与速度和加速度相关的配置值，使其能够接受双精度浮点数。
- 修改了 `AxisControllerEnums.h` 中 `MAX_SPEED` 的定义，以匹配新的数值类型。
- 更新了 `MutiCamApp.cpp` 中与速度和加速度相关的槽函数签名和内部逻辑，以处理 `double` 类型。
- 调整了 `MutiCamApp.ui` 文件中涉及速度和加速度的 `QSpinBox` 为 `QDoubleSpinBox`，并更新了相应的属性设置，如 `decimals` 和 `singleStep`。

+++ src/core/MutiCamApp.h
@@ -243,10 +243,10 @@ private slots:
     void onEnableZChanged(bool enabled);   // Z轴使能状态改变
     
     // 速度和加速度设置槽函数
-    void onSpeedChanged(int speed);        // 速度改变
-    void onAccelChanged(int accel);        // 加速度改变
-    void onMaxSpeedChanged(int maxSpeed);  // 最大速度限制改变  
-    void onMaxAccelChanged(int maxAccel);  // 最大加速度限制改变
+    void onSpeedChanged(double speed);        // 速度改变
+    void onAccelChanged(double accel);        // 加速度改变
+    void onMaxSpeedChanged(double maxSpeed);  // 最大速度限制改变
+    void onMaxAccelChanged(double maxAccel);  // 最大加速度限制改变
     
     // 运动模式切换槽函数
     void onMotionModeChanged(int mode);    // 运动模式改变（0=点动，1=连续）
